move file to own folder
